Moving the util and updating references in tests
